requirements-dev.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
numpy>=1.21.0
opencv-python>=4.5.0
moviepy>=1.0.0
pydub>=0.25.1
faster-whisper>=0.6.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0

# Testing
pytest-mock>=3.10.0

# Type Hints
typing-extensions>=4.0.0
types-python-dateutil>=2.8.0
types-requests>=2.26.0
